requirements.txt
Django==2.2.5
psycopg2
djangorestframework == 3.10.3
django-cors-headers==3.2.0
markdown
beautifulsoup4==4.8.1
requests==2.22.0
amqp